pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "holocene"
version = "0.1.0"
description = "Your personal geological record of the present - activity tracking and AI assistant"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Arthur", email = "endar@users.noreply.github.com"}
]
keywords = ["productivity", "tracking", "llm", "assistant", "adhd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "pypdf>=3.17.0",
    "pdfplumber>=0.10.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pyusb>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
holo = "holocene.cli.main:cli"

[tool.setuptools]
packages = ["holocene"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
